Track modified files in this validation run to filter diagnostics published
