+++ Code/Language/m/src/m/validation/global/GlobalData.java
@@ -1,8 +1,10 @@
 package m.validation.global;
 
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import m.library.Library;
 import m.model.Game;
@@ -12,10 +14,12 @@ public class GlobalData
 {
 	public final Game game;
 	public final Map<String,List<Problem>> problems;
+	public Set<String> modifiedFiles;
 
 	public GlobalData()
 	{
 		this.game = new Game(Library.ENGLISH);
 		this.problems = new HashMap<>();
+		this.modifiedFiles = new HashSet<>();
 	}
 }
